API now generates the rig.conf and wallet.conf according to the rig config in database. added routes accordinaly to the frontend

+++ resources/flight_sheet.py
@@ -61,6 +61,11 @@ class FlightSheet(Resource):
     @jwt_required()
     def delete(self):
         fs = request.args['fs']
+        is_linked = db.config.find_one({
+            "flightsheet": fs
+        })
+        if is_linked is not None:
+            return {"error": "This flightsheet"}, 403
         db.flightsheets.delete_one({
             "user_id": current_identity.id,
             "_id": ObjectId(fs)
